docs: update README and finalize class adjustments
- Fix class diagram: remove damageLevel from JogoFisico
- Update JogoFisico description to reflect damage fee tracking
- Update project structure listing all current source files
- Fix terminal run command to use Main class
- Remove Tomaz from authors list
- Add toString() to Jogo class
- Fix Jogo, JogoFisico, JogoDigital constructors to use double for valorDiario
- Switch JogoDigital.isDisponivel() from >= back to >
- Add newline to Cliente.toString()
- Refactor Locacao formatting and availability check
- Fix Main.java unavailable stock test to properly exhaust units

+++ GameRent/src/GameRent/JogoDigital.java
@@ -6,7 +6,7 @@ public class JogoDigital extends Jogo {
 	private int limiteAcessosSimultaneos = 5; // Exemplo
 	private int acessosAtivos = 0;
 
-	public JogoDigital(String nome, Plataforma plataforma, String genero, String chaveAcesso, int valorDiario, ClassificacaoEtaria classificacao, int tamanhoGB){
+	public JogoDigital(String nome, Plataforma plataforma, String genero, String chaveAcesso, double valorDiario, ClassificacaoEtaria classificacao, int tamanhoGB){
 		super(nome,plataforma,genero,valorDiario,classificacao);
 
 		this.chaveAcesso = chaveAcesso;
