+++ GameRent/src/GameRent/JogoFisico.java
@@ -4,7 +4,7 @@ public class JogoFisico extends Jogo{
 	private int qtdEstoque;
 	private int qtdAlugado = 0;
 
-	public JogoFisico(String nome, Plataforma plataforma, String genero, int qtdEstoque, int valorDiario, ClassificacaoEtaria classificacao){
+	public JogoFisico(String nome, Plataforma plataforma, String genero, int qtdEstoque, double valorDiario, ClassificacaoEtaria classificacao){
 		super(nome,plataforma,genero,valorDiario,classificacao);
 
 		this.qtdEstoque = qtdEstoque;
@@ -40,6 +40,4 @@ public class JogoFisico extends Jogo{
 	public int getQntdEstoque() {
 		return qtdEstoque;
 	}
-
-
 }
